feat: migrate backend to TypeScript with express
